all files: rename evse.cpp/evse.h to main.cpp/main.h
