+++ SmartEVSE-3/src/OneWire.cpp
@@ -27,7 +27,7 @@
 #include <string.h>
 #include <Preferences.h>
 
-#include "evse.h"
+#include "main.h"
 #include "utils.h"
 #include "OneWire.h"
 #include "OneWireESP32.h"
