+++ SmartEVSE-3/src/utils.cpp
@@ -25,7 +25,7 @@
 #include <stdio.h>
 #include <stdlib.h>
 
-#include "evse.h"
+#include "main.h"
 #include "utils.h"
 
 unsigned long pow_10[10] = {1, 10, 100, 1000, 10000, 100000, 1000000, 10000000, 100000000, 1000000000};
